rename tests to `render_test.rs`
